bugfix: do not hide the table if settings are wrong (see below)
wrong settings: until now the user was able to resize the columns when the gallery was active!
therefore every column (should have been) was invisible and that's what was stored as settings.
settings = every column hidden = no way to display the table again...
fix: 1. do not save hide-every-column-settings. 2. display the default settings if everything was insvisible